refactored Objects page to use standard page format
