+++ data/descriptor-formats/formats.json
@@ -65,6 +65,7 @@
   "_material": [["pick", "materialObject"]],
   "_misc": [["pick", "misc"]],
   "_quality": [["pick", "qualityObject"]],
+  "_a(n)quality": [["a(n)"], ["pick", "qualityObject"]],
   "_treasure": [["pick", "treasure"]],
   "_weapon": [["pick", "weapon"]],
   "_writing": [["pick", "writing"]],
